% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/STATegRa_omicsCLUST_bioDist.R
\name{bioDistclass}
\alias{bioDistclass}
\title{bioDistclass}
\usage{
bioDistclass(name, distance, map.name, map.metadata, params)
}
\arguments{
\item{name}{Name assigned to the object}

\item{distance}{Matrix giving the distance between features}

\item{map.name}{Charactering giving the name of the bioMap object used to compute the distance}

\item{map.metadata}{List of parameters used to generate the mapping}

\item{params}{List of parameters used to generate the distance}
}
\description{
Class to manage mappings between genomic features.
}

