% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/STATegRa_omicsPCA_plotting.R
\name{biplotRes}
\alias{biplotRes}
\alias{biplotRes,caClass,character,numeric,character-method}
\title{Biplot of component analysis}
\usage{
biplotRes(object, type, comps, block, title=NULL, colorCol=NULL,
                 sizeValues=c(2, 4), shapeValues=c(17, 0), background=TRUE,
                 pointSize=4, labelSize=NULL, axisSize=NULL, titleSize=NULL)
}
\arguments{
\item{object}{\code{caClass} object containing component analysis results}

\item{type}{Character specifying which components to plot; "common", "individual" or "both"}

\item{comps}{Components to plot. If \code{combined=FALSE}, specifies the component indices to use as x and y for the plot. Otherwise, the component from the first block and the component from second block to plot together.}

\item{block}{Which block to plot, either "1" or "2" or the name of the block.}

\item{title}{Plot title}

\item{colorCol}{Character specifying a pData column to use to colorise the plot points}

\item{sizeValues}{Vector containing sizes for scores and loadings}

\item{shapeValues}{Vector indicating the shapes for scores and loadings}

\item{background}{Logical, whether to use a grey background}

\item{pointSize}{Size of plot points}

\item{labelSize}{Size of plot labels if not NULL}

\item{axisSize}{Size of axis text}

\item{titleSize}{Size of title text}
}
\value{
\code{ggplot2} object
}
\description{
Generate a biplot of component analysis results
}
\examples{
data("STATegRa_S3")
B1 <- createOmicsExpressionSet(Data=Block1.PCA,pData=ed.PCA,
                               pDataDescr=c("classname"))
B2 <- createOmicsExpressionSet(Data=Block2.PCA,
                               pData=ed.PCA,pDataDescr=c("classname"))
# Omics components analysis
discoRes <- omicsCompAnalysis(Input=list(B1,B2),Names=c("expr","mirna"),
                              method="DISCOSCA",Rcommon=2,Rspecific=c(2,2),
                              center=TRUE,scale=TRUE,weight=TRUE)
jiveRes <- omicsCompAnalysis(Input=list(B1,B2),Names=c("expr","mirna"),
                             method="JIVE",Rcommon=2,Rspecific=c(2,2),
                             center=TRUE,scale=TRUE,weight=TRUE)

o2plsRes <- omicsCompAnalysis(Input=list(B1,B2),Names=c("expr","mirna"),
                              method="O2PLS",Rcommon=2,Rspecific=c(2,2),
                              center=TRUE,scale=TRUE,weight=TRUE)

# Biplot common part. DISCO-SCA

biplotRes(object=discoRes,type="common",comps=c(1,2),block="",
          title=NULL,colorCol="classname",sizeValues=c(2,4),
          shapeValues=c(17,0),background=TRUE,pointSize=4,
          labelSize=NULL,axisSize=NULL,titleSize=NULL)

# Biplot common part. O2PLS

p1 <- biplotRes(object=o2plsRes,type="common",comps=c(1,2),
                block="expr",title=NULL,colorCol="classname",
                sizeValues=c(2,4),shapeValues=c(17,0),
                background=TRUE,pointSize=4,labelSize=NULL,
                axisSize=NULL,titleSize=NULL)
p2 <- biplotRes(object=o2plsRes,type="common",comps=c(1,2),
                block="mirna",title=NULL,colorCol="classname",
                sizeValues=c(2,4),shapeValues=c(17,0),
                background=TRUE,pointSize=4,labelSize=NULL,
                axisSize=NULL,titleSize=NULL)

# Biplot distinctive part. O2PLS

p1 <- biplotRes(object=discoRes,type="individual",comps=c(1,2),
                block="expr",title=NULL,colorCol="classname",
                sizeValues=c(2,4),shapeValues=c(17,0),
                background=TRUE,pointSize=4,labelSize=NULL,
                axisSize=NULL,titleSize=NULL)
p2 <- biplotRes(object=discoRes,type="individual",comps=c(1,2),
                block="mirna",title=NULL,colorCol="classname",
                sizeValues=c(2,4),shapeValues=c(17,0),
                background=TRUE,pointSize=4,labelSize=NULL,
                axisSize=NULL,titleSize=NULL)
}
\author{
Patricia Sebastian-Leon
}

