% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/STATegRa_omicsPCA_caClass.R
\docType{class}
\name{caClass-class}
\alias{caClass-class}
\title{caClass}
\description{
Stores the results of any of the omicsPCA analyses.
}
\section{Slots}{

\describe{
\item{\code{InitialData}}{List of ExpressionSets, one for each set of omics data}

\item{\code{Names}}{Character vector giving names for the input data}

\item{\code{preprocessing}}{Character vector describing the preprocessing applied to the data}

\item{\code{preproData}}{List of matrices containing data after preprocessing}

\item{\code{caMethod}}{Character giving the component analysis method name}

\item{\code{commonComps}}{Numeric giving the number of common components}

\item{\code{distComps}}{Numeric vector giving the number of distinctive components for each block}

\item{\code{scores}}{List of matrices of common and distinctive scores}

\item{\code{loadings}}{List of matrices of common and distinctive loadings}

\item{\code{VAF}}{List of matrices indicating VAF (Variability Explained For) for each component in each block of data}

\item{\code{others}}{List containing other miscellaneous information specific to different SCA methods}
}}
\author{
Patricia Sebastian Leon
}

