% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/STATegRa_omicsPCA_methods.R
\name{selectCommonComps}
\alias{selectCommonComps}
\alias{selectCommonComps,matrix,matrix,numeric-method}
\title{Select common components in two data blocks}
\usage{
selectCommonComps(X, Y, Rmax)
}
\arguments{
\item{X}{Matrix of omics data}

\item{Y}{Matrix of omics data}

\item{Rmax}{Maximum number of common components to find}
}
\value{
A list with components:
\describe{
     \item{common}{Optimal number of common components}
     \item{ssqs}{Matrix of SSQ for each block and estimator}
     \item{pssq}{\code{\link{ggplot}} object showing SSQ for each block and estimator}
     \item{pratios}{\code{\link{ggplot}} object showing SSQ ratios between each block and estimator}
}
}
\description{
This function applies a Simultaneous Component Analysis (SCA). The idea is that the scores for both blocks should have a similar behaviour if the components are in the common mode. Evaluation is by the ratios between the explained variances (SSQ) of each block and the estimator. The highest component count with 0.8 < ratio < 1.5 is selected.
}
\examples{
data(STATegRa_S3)
cc <- selectCommonComps(X=Block1.PCA, Y=Block2.PCA, Rmax=3)
cc$common
cc$pssq
cc$pratios
}
\author{
Patricia Sebastian-Leon
}

