% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STATegRa_package.R
\name{STATegRa-defunct}
\alias{STATegRa-defunct}
\title{Defunct functions in STATegRa}
\description{
These functions have are defunct and no longer available
}
\details{
\itemize{
\item{holistOmics: replaced by \code{\link{omicsNPC}}}
}
}

