% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STATegRa_package.R
\name{STATegRa-deprecated}
\alias{STATegRa-deprecated}
\title{Deprecated functions in STATegRa}
\description{
These functions have been deprecated in STATegRa
}
\details{
\itemize{
\item{holistOmics: \code{\link{omicsNPC}}}
}
}

