% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STATegRa_package.R
\docType{data}
\name{STATegRa_data_TCGA_BRCA}
\alias{STATegRa_data_TCGA_BRCA}
\alias{TCGA_BRCA_Data}
\title{STATegRa data}
\format{One list, which contains three ExpressionSet objects.}
\source{
See \url{https://tcga-data.nci.nih.gov/tcga/}.
}
\description{
Data were downloaded from TCGA data portal, \url{https://tcga-data.nci.nih.gov/tcga/}.
We downloaded sixteen tumour samples and the sixteen matching normal, for Breast invasive carcinoma, BRCA, batch 93.
Herein, three types of data modalities are included, RNAseq (\code{TCGA_BRCA_Data$RNAseq}), RNAseqV2 (\code{TCGA_BRCA_Data$RNAseqV2})
and Expression-Genes (\code{TCGA_BRCA_Data$Microarray}). The Data Level was set to Level 3.
For each data type, we pooled all data to one matrix, where rows corresponded to genes and columns to samples.
Only the first 100 genes are included.
}
\examples{
# load data
data(TCGA_BRCA_Batch_93)
}
\author{
Nestoras Karathanasis, Vincenzo Lagani
}
