\alias{STATegRa}
\alias{STATegRa-package}
\docType{package}
\name{STATegRa}
\title{STATegRa}
\description{
STATegRa is a package for the integrative analysis of multi-omic data-sets.

For full information, see the user's guide.
}
\seealso{\code{\link{STATegRaUsersGuide}}}
\examples{\dontrun{STATegRaUsersGuide()}}