% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/STATegRa_omicsPCA_methods.R
\name{modelSelection}
\alias{modelSelection}
\alias{modelSelection,list,numeric,character-method}
\title{Find optimal common and distinctive components}
\usage{
modelSelection(Input, Rmax, fac.sel, varthreshold=NULL, nvar=NULL, PCnum=NULL)
}
\arguments{
\item{Input}{List of two ExpressionSet objects}

\item{Rmax}{Maximum common components (see \code{\link{selectCommonComps}})}

\item{fac.sel}{PCA criteria (see \code{\link{PCA.selection}})}

\item{varthreshold}{Cumulative variance criteria (see \code{\link{PCA.selection}})}

\item{nvar}{Relative variance criteria (see \code{\link{PCA.selection}})}

\item{PCnum}{Fixed component number (see \code{\link{PCA.selection}})}
}
\value{
List containing:
\describe{
     \item{common}{Number of common components}
     \item{dist}{Number of distinct components per input block}
}
}
\description{
Uses \code{\link{selectCommonComps}} and \code{\link{PCA.selection}} to estimate the optimal number of common and distinctive components according to given selection criteria.
}
\examples{
data(STATegRa_S3)
B1 <- createOmicsExpressionSet(Data=Block1.PCA,pData=ed.PCA,pDataDescr=c("classname"))
B2 <- createOmicsExpressionSet(Data=Block2.PCA,pData=ed.PCA,pDataDescr=c("classname"))
ms <- modelSelection(Input=list(B1, B2), Rmax=4, fac.sel="single\\\%", varthreshold=0.03)
ms
}
\author{
Patricia Sebastian-Leon
}
\seealso{
\code{\link{selectCommonComps}},\code{\link{PCA.selection}},\code{\link{omicsCompAnalysis}}
}

