% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/STATegRa_package.R
\docType{package}
\name{STATegRa}
\alias{STATegRa}
\alias{STATegRa-package}
\title{STATegRa}
\description{
STATegRa is a package for the integrative analysis of multi-omic data-sets.

For full information, see the user's guide.
}
\seealso{
\code{\link{STATegRaUsersGuide}}
}

