% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/STATegRa_fused.R
\name{STATegRaUsersGuide}
\alias{STATegRaUsersGuide}
\title{STATegRaUsersGuide}
\usage{
STATegRaUsersGuide(view = TRUE)
}
\arguments{
\item{view}{Whether to open a browser}
}
\value{
The path to the documentation
}
\description{
Finds the location of the STATegRa User's Guide and optionally opens it.
}
\examples{
STATegRaUsersGuide(view=FALSE)
}
\author{
David Gomez-Cabrero
}

